# app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "API documentation for E-Commerce Flask application"
  },
  "servers": [{"url": "http://localhost:5000"}],
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "email": {"type": "string"},
          "full_name": {"type": "string"},
          "role": {"type": "string"},
          "is_verified": {"type": "boolean"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string"}
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "sku": {"type": "string"},
          "description": {"type": "string"},
          "price": {"type": "number"},
          "stock": {"type": "integer"},
          "image_url": {"type": "string"},
          "category": {"type": "string"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string"},
          "updated_at": {"type": "string"}
        }
      },
      "ProductVariant": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "product_id": {"type": "string"},
          "sku": {"type": "string"},
          "name": {"type": "string"},
          "color": {"type": "string"},
          "size": {"type": "string"},
          "price": {"type": "number"},
          "stock": {"type": "integer"},
          "image_url": {"type": "string"},
          "is_active": {"type": "boolean"}
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "slug": {"type": "string"},
          "description": {"type": "string"},
          "image_url": {"type": "string"},
          "parent_id": {"type": "string"},
          "sort_order": {"type": "integer"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string"},
          "updated_at": {"type": "string"}
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "example": "user@example.com"},
                  "password": {"type": "string", "example": "password123"},
                  "full_name": {"type": "string", "example": "John Doe"}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "Registered, OTP sent"}}
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify OTP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "code"],
                "properties": {
                  "email": {"type": "string", "example": "user@example.com"},
                  "code": {"type": "string", "example": "123456"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Verified"}}
      }
    },
    "/api/auth/resend-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Resend OTP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {"type": "string", "example": "user@example.com"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "OTP sent"}}
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "example": "user@example.com"},
                  "password": {"type": "string", "example": "password123"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Login successful"}}
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh token",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Token refreshed"}}
      }
    },
    "/api/dashboard/overview": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get dashboard overview (Admin only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dashboard data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "overview": {
                      "type": "object",
                      "properties": {
                        "total_revenue": {"type": "number"},
                        "today_revenue": {"type": "number"},
                        "today_orders": {"type": "integer"},
                        "total_orders": {"type": "integer"},
                        "total_users": {"type": "integer"}
                      }
                    },
                    "top_products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {"type": "string"},
                          "name": {"type": "string"},
                          "price": {"type": "number"},
                          "total_sold": {"type": "integer"}
                        }
                      }
                    },
                    "revenue_chart": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {"type": "string"},
                          "revenue": {"type": "number"}
                        }
                      }
                    },
                    "encrypted": {"type": "boolean"}
                  }
                }
              }
            }
          },
          "403": {"description": "Admin access required"}
        }
      }
    },
    "/api/decrypt/test": {
      "post": {
        "tags": ["Decrypt Helper"],
        "summary": "Decrypt single value",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["encrypted_value"],
                "properties": {
                  "encrypted_value": {"type": "string", "example": "gAAAAABh5K8_..."}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Decrypted value"}}
      }
    },
    "/api/decrypt/response": {
      "post": {
        "tags": ["Decrypt Helper"],
        "summary": "Decrypt entire response",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["response"],
                "properties": {
                  "response": {"type": "object", "description": "Encrypted response object"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Decrypted response"}}
      }
    },
    "/api/users": {
      "get": {
        "tags": ["User Management"],
        "summary": "Get all users (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/User"}
                    },
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "per_page": {"type": "integer"},
                    "total_pages": {"type": "integer"}
                  }
                }
              }
            }
          },
          "403": {"description": "Admin access required"}
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": ["User Management"],
        "summary": "Get user by ID (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "404": {"description": "User not found"}
        }
      }
    },
    "/api/users/{user_id}/toggle-status": {
      "put": {
        "tags": ["User Management"],
        "summary": "Lock/Unlock user account (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {"description": "User status toggled"},
          "400": {"description": "Cannot lock/unlock own account"},
          "404": {"description": "User not found"}
        }
      }
    },
    "/api/users/{user_id}/role": {
      "put": {
        "tags": ["User Management"],
        "summary": "Update user role (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": ["ADMIN", "USER"],
                    "example": "ADMIN"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Role updated"},
          "400": {"description": "Invalid role or cannot change own role"},
          "404": {"description": "User not found"}
        }
      }
    },
    "/api/users/{user_id}/reset-password": {
      "put": {
        "tags": ["User Management"],
        "summary": "Reset user password (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["new_password"],
                "properties": {
                  "new_password": {
                    "type": "string",
                    "example": "newpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Password reset successfully"},
          "400": {"description": "Invalid password"},
          "404": {"description": "User not found"}
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Get all products (Public)",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 20}},
          {"name": "category", "in": "query", "schema": {"type": "string"}},
          {"name": "is_active", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {"type": "array", "items": {"$ref": "#/components/schemas/Product"}},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "per_page": {"type": "integer"},
                    "total_pages": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/add": {
      "post": {
        "tags": ["Product Management"],
        "summary": "Create product (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": {"type": "string", "example": "iPhone 15 Pro"},
                  "sku": {"type": "string", "example": "IP15PRO-BLK"},
                  "description": {"type": "string"},
                  "price": {"type": "number", "example": 999.99},
                  "stock": {"type": "integer", "example": 100},
                  "image_url": {"type": "string"},
                  "category": {"type": "string", "example": "Electronics"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Product created"},
          "400": {"description": "Validation error"}
        }
      }
    },
    "/api/products/{product_id}": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Get product by ID (Public)",
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Product details"},
          "404": {"description": "Product not found"}
        }
      },
      "put": {
        "tags": ["Product Management"],
        "summary": "Update product (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "sku": {"type": "string"},
                  "description": {"type": "string"},
                  "price": {"type": "number"},
                  "stock": {"type": "integer"},
                  "category": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Product updated"},
          "404": {"description": "Product not found"}
        }
      },
      "delete": {
        "tags": ["Product Management"],
        "summary": "Delete product (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Product deleted"},
          "404": {"description": "Product not found"}
        }
      }
    },
    "/api/products/{product_id}/toggle-status": {
      "put": {
        "tags": ["Product Management"],
        "summary": "Toggle product visibility (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Status toggled"}
        }
      }
    },
    "/api/products/{product_id}/stock": {
      "put": {
        "tags": ["Product Management"],
        "summary": "Update product stock (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["stock"],
                "properties": {
                  "stock": {"type": "integer", "example": 50}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Stock updated"}
        }
      }
    },
    "/api/products/{product_id}/images": {
      "post": {
        "tags": ["Product Management"],
        "summary": "Upload product image (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {"type": "string", "format": "binary"},
                  "is_primary": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Image uploaded"}
        }
      }
    },
    "/api/products/{product_id}/variants": {
      "post": {
        "tags": ["Product Management"],
        "summary": "Create product variant (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sku", "name"],
                "properties": {
                  "sku": {"type": "string", "example": "IP15PRO-BLK-256"},
                  "name": {"type": "string", "example": "Black - 256GB"},
                  "color": {"type": "string", "example": "Black"},
                  "size": {"type": "string", "example": "256GB"},
                  "price": {"type": "number", "example": 1099.99},
                  "stock": {"type": "integer", "example": 50}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Variant created"}
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Category Management"],
        "summary": "Get all categories (Public)",
        "parameters": [
          {"name": "include_children", "in": "query", "schema": {"type": "boolean", "default": false}},
          {"name": "only_parents", "in": "query", "schema": {"type": "boolean", "default": false}},
          {"name": "is_active", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {"type": "array", "items": {"$ref": "#/components/schemas/Category"}},
                    "total": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/tree": {
      "get": {
        "tags": ["Category Management"],
        "summary": "Get category tree with hierarchy (Public)",
        "responses": {
          "200": {
            "description": "Category tree"
          }
        }
      }
    },
    "/api/categories/add": {
      "post": {
        "tags": ["Category Management"],
        "summary": "Create category (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slug"],
                "properties": {
                  "name": {"type": "string", "example": "Electronics"},
                  "slug": {"type": "string", "example": "electronics"},
                  "description": {"type": "string"},
                  "image_url": {"type": "string"},
                  "parent_id": {"type": "string"},
                  "sort_order": {"type": "integer", "example": 1}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Category created"},
          "400": {"description": "Validation error"}
        }
      }
    },
    "/api/categories/{category_id}": {
      "get": {
        "tags": ["Category Management"],
        "summary": "Get category by ID (Public)",
        "parameters": [
          {"name": "category_id", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "include_children", "in": "query", "schema": {"type": "boolean", "default": true}}
        ],
        "responses": {
          "200": {"description": "Category details"},
          "404": {"description": "Category not found"}
        }
      },
      "put": {
        "tags": ["Category Management"],
        "summary": "Update category (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "category_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "slug": {"type": "string"},
                  "description": {"type": "string"},
                  "parent_id": {"type": "string"},
                  "sort_order": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Category updated"},
          "404": {"description": "Category not found"}
        }
      },
      "delete": {
        "tags": ["Category Management"],
        "summary": "Delete category (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "category_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Category deleted"},
          "400": {"description": "Cannot delete category with children"},
          "404": {"description": "Category not found"}
        }
      }
    },
    "/api/categories/slug/{slug}": {
      "get": {
        "tags": ["Category Management"],
        "summary": "Get category by slug (Public)",
        "parameters": [
          {"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "include_children", "in": "query", "schema": {"type": "boolean", "default": true}}
        ],
        "responses": {
          "200": {"description": "Category details"},
          "404": {"description": "Category not found"}
        }
      }
    },
    "/api/categories/{category_id}/toggle-status": {
      "put": {
        "tags": ["Category Management"],
        "summary": "Toggle category visibility (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "category_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Status toggled"}
        }
      }
    }
  }
}